+++ delphi-frontend/src/main/java/org/sonar/plugins/communitydelphi/api/ast/AnonymousMethodNode.java
@@ -41,4 +41,8 @@ public interface AnonymousMethodNode extends ExpressionNode {
   boolean isFunction();
 
   boolean isProcedure();
+
+  CompoundStatementNode getStatementBlock();
+
+  boolean isEmpty();
 }